pyproject.toml
[tool.poetry]
name = "image_proxy"
version = "0.1.0"
description = ""
authors = ["Eldor Pulatov <psanti99.sp@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
djangorestframework = "^3.12.2"
Pillow = "^8.0.1"
requests = "^2.25.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
